TODO: do death checker like meal checker
