+++ philo/error.c
@@ -1,0 +1,8 @@
+/* 27/6/2023 14:12:00
+* Sin sanitice thread
+	? Funciona el meal checker
+
+! Con sanitice thread
+	? Funcina el meal checker, pero sigue imprimiendo las acciones del
+	? ultimo hilo.
+*/
