+++ philo/srcs/main.c
@@ -6,11 +6,24 @@
 /*   By: psegura- <psegura-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/05 12:11:11 by psegura-          #+#    #+#             */
-/*   Updated: 2023/06/24 03:20:47 by psegura-         ###   ########.fr       */
+/*   Updated: 2023/06/27 13:58:22 by psegura-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philo.h"
+
+int init_checkers(t_data *c)
+{
+	pthread_t	meals_check;
+
+	if (c->args[MEALS_C] > 0)
+	{
+		pthread_create(&meals_check, NULL, &meals_checker, (void *)c);
+		pthread_detach(meals_check);
+		pthread_join(meals_check, NULL);
+	}
+	return (0);
+}
 
 int	main(int argc, char **argv)
 {
@@ -25,6 +38,7 @@ int	main(int argc, char **argv)
 		{
 			init_mutex(&c);
 			init_philosophers(&c);
+			init_checkers(&c);
 			pthread_mutex_lock(&c.death);
 		}
 	}
