+++ philo/srcs/utils.c
@@ -6,7 +6,7 @@
 /*   By: psegura- <psegura-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/05 13:02:02 by psegura-          #+#    #+#             */
-/*   Updated: 2023/06/26 16:07:31 by psegura-         ###   ########.fr       */
+/*   Updated: 2023/06/27 14:06:24 by psegura-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -72,6 +72,7 @@ int	ft_strlen(char *str)
 
 int	print_game(t_philo *phils, char *str, int lock)
 {
+	(void)lock;
 	pthread_mutex_lock(&phils->c->printing);
 	if (is_dead(phils) == FALSE)
 		printf(PRINTER, time_dif(phils->c->program_start_time), phils->id, str);
